[release prep] Fix warnings in the SDKSampleApp

+++ SDKSampleApp/SDKSampleApp/Screens/Transaction/RefundViewController.m
@@ -56,7 +56,7 @@
 
 -(NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
-    int numTransactions = [self.transactionRecords count];
+    NSUInteger numTransactions = [self.transactionRecords count];
     if (numTransactions == 0) {
         numTransactions += 1;
         _showingNoneAvailable = YES;
